layout/style/test/test_bug200089.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=200089
-->
<head>
  <title>Test for Bug 200089</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=200089">Mozilla Bug 200089</a>
<div id="display" style="width: 600px">
  <table border="0" id="t" style="width: 300px; margin-left: auto; margin-right: auto">
    <tr><td>Cell</td></tr>
  </table>
</div>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 200089 **/
is(getComputedStyle($("t"), "").width, "300px",
   "Used width should match specified width in this case");
</script>
</pre>
</body>
</html>
